backend 0.1

+++ backend/src/controllers/TaskController.ts
@@ -4,17 +4,21 @@ import { checkValidationErrors } from "../utils/helpers";
 import TaskService from "../services/TaskService";
 import SessionService from "../services/SessionService";
 
+import User from "../models/User";
+import Task from "../models/Task";
+
+import APIError from "../exceptions/api-error";
+import UserService from "../services/UserService";
+
 
 class TaskController {
     async create(req: Request, res: Response, next: NextFunction){
         try {
             // validation
             await checkValidationErrors(req);
-            // await TaskService.isValidPriorityType(req.body.priority);
-            req.body.expiresAt = new Date();
             
             // making auth user be author by default
-            req.body.author = (await SessionService.verify(req.cookies["sessionId"])).id;
+            req.body.creatorId = String(await SessionService.verify(req.cookies["sessionId"]));
 
             // creating task
             const task = await TaskService.add(req.body);
@@ -26,7 +30,9 @@ class TaskController {
     }
     async getAll(req: Request, res: Response, next: NextFunction){
         try {
-            const tasks = await TaskService.getList();
+            // getting current user
+            const currentUserId = String(await SessionService.verify(req.cookies["sessionId"]));
+            const tasks = await TaskService.getList(currentUserId, String(req.query.time), Boolean(req.query.groupByResponsible));
 
             res.send(tasks);
         } catch(err) {
@@ -43,7 +49,20 @@ class TaskController {
     }
     async patchOne(req: Request, res: Response, next: NextFunction){
         try {
-            const task = await TaskService.patchOne(req.params.id, req.body);
+            const userId = String(await SessionService.verify(req.cookies["sessionId"]));
+            const taskCandidate = await TaskService.getOne(req.params.id);
+            let hasFullPermission = false;
+            
+            if (userId === taskCandidate.creatorId) {
+                hasFullPermission = true;
+            }
+            if (userId !== taskCandidate.creatorId && userId !== taskCandidate.responsibleId)
+            {
+                throw APIError.UnauthorizedError('not enough permission');
+            }
+
+            const task = await TaskService.patchOne(req.params.id, req.body, hasFullPermission);
+
             res.send(task);          
         } catch(err) {
             next(err);            
@@ -51,8 +70,13 @@ class TaskController {
     }
     async deleteOne(req: Request, res: Response, next: NextFunction){
         try {
-            await TaskService.deleteOne(req.params.id);
+            const userId = String(await SessionService.verify(req.cookies["sessionId"]));
+            const taskCandidate = await TaskService.getOne(req.params.id);
+            if (userId !== taskCandidate.creatorId) {
+                throw APIError.UnauthorizedError('not enough permission');
+            }
 
+            await TaskService.deleteOne(req.params.id);
             res.status(204).send();
         } catch(err) {
             next(err);
